fix(typedef name)
