+++ sf_car/src/stm32f7xx_it.c
@@ -1,0 +1,42 @@
+/**
+  ******************************************************************************
+  * @file    stm32f7xx_it.c
+  * @author  Ac6
+  * @version V1.0
+  * @date    02-Feb-2015
+  * @brief   Default Interrupt Service Routines.
+  ******************************************************************************
+*/
+
+/* Includes ------------------------------------------------------------------*/
+#include "stm32f7xx_hal.h"
+#include "stm32f7xx.h"
+#ifdef USE_RTOS_SYSTICK
+#include <cmsis_os.h>
+#endif
+#include "stm32f7xx_it.h"
+
+/* Private typedef -----------------------------------------------------------*/
+/* Private define ------------------------------------------------------------*/
+/* Private macro -------------------------------------------------------------*/
+/* Private variables ---------------------------------------------------------*/
+/* Private function prototypes -----------------------------------------------*/
+/* Private functions ---------------------------------------------------------*/
+
+/******************************************************************************/
+/*            	  	    Processor Exceptions Handlers                         */
+/******************************************************************************/
+
+/**
+  * @brief  This function handles SysTick Handler, but only if no RTOS defines it.
+  * @param  None
+  * @retval None
+  */
+void SysTick_Handler(void)
+{
+	HAL_IncTick();
+	HAL_SYSTICK_IRQHandler();
+#ifdef USE_RTOS_SYSTICK
+	osSystickHandler();
+#endif
+}
