--- src/main.c
@@ -1,163 +1,0 @@
-#include "stm32f7xx.h"
-#include "stm32746g_discovery.h"
-#include "stm32746g_discovery_lcd.h"
-#include <string.h>
-#include <stdlib.h>
-
-UART_HandleTypeDef uart_handle; //uart typedef
-TIM_HandleTypeDef tim_handle; 	//pwm timer typedef
-TIM_OC_InitTypeDef sConfig;		// pwm typedef
-GPIO_InitTypeDef servo_motor;
-
-//#undef __GNUC__
-
-#ifdef __GNUC__
-#define PUTCHAR_PROTOTYPE int __io_putchar(int ch)
-#else
-#define PUTCHAR_PROTOTYPE int fputc(int ch, FILE *f)
-#endif /* __GNUC__ */
-
-static void SystemClock_Config(void);
-static void Error_Handler(void);
-static void MPU_Config(void);
-static void CPU_CACHE_Enable(void);
-
-void uart_setup() {
-	//uart setup
-	uart_handle.Init.BaudRate = 115200;
-	uart_handle.Init.WordLength = UART_WORDLENGTH_8B;
-	uart_handle.Init.StopBits = UART_STOPBITS_1;
-	uart_handle.Init.Parity = UART_PARITY_NONE;
-	uart_handle.Init.HwFlowCtl = UART_HWCONTROL_NONE;
-	uart_handle.Init.Mode = UART_MODE_TX_RX;
-	BSP_COM_Init(COM1, &uart_handle);
-
-}
-
-void servo_motor_setup() {
-
-	__HAL_RCC_GPIOB_CLK_ENABLE();
-
-	//cooler setup
-	cooler.Pin = GPIO_PIN_4;
-	cooler.Mode = GPIO_MODE_AF_PP;
-	cooler.Pull = GPIO_NOPULL;
-	cooler.Speed = GPIO_SPEED_HIGH;
-	cooler.Alternate = GPIO_AF2_TIM3;
-	HAL_GPIO_Init(GPIOB, &servo_motor);
-
-}
-
-void pwm_setup() {
-
-	__HAL_RCC_TIM3_CLK_ENABLE();
-
-	//pwm setup
-	tim_handle.Instance = TIM3;
-	tim_handle.Init.Period = 100;
-	tim_handle.Init.Prescaler = 1;
-	tim_handle.Init.ClockDivision = TIM_CLOCKDIVISION_DIV1;
-	tim_handle.Init.CounterMode = TIM_COUNTERMODE_UP;
-	HAL_TIM_Base_Init(&tim_handle);
-	HAL_TIM_Base_Start_IT(&tim_handle);
-
-	//pwm config
-	sConfig.OCMode = TIM_OCMODE_PWM1;
-	sConfig.Pulse = 0;
-
-	//pwm init & start
-	HAL_TIM_PWM_ConfigChannel(&tim_handle, &sConfig, TIM_CHANNEL_1);
-	HAL_TIM_PWM_Init(&tim_handle);
-	HAL_TIM_PWM_Start(&tim_handle, TIM_CHANNEL_1);
-
-}
-
-int main(void) {
-
-	MPU_Config();
-	CPU_CACHE_Enable();
-	HAL_Init();
-	SystemClock_Config();
-
-	uart_setup();
-	pwm_setup();
-
-	while (1) {
-	}
-}
-
-
-PUTCHAR_PROTOTYPE {
-	HAL_UART_Transmit(&uart_handle, (uint8_t *) &ch, 1, 0xFFFF);
-
-	return ch;
-}
-static void SystemClock_Config(void) {
-	RCC_ClkInitTypeDef RCC_ClkInitStruct;
-	RCC_OscInitTypeDef RCC_OscInitStruct;
-
-	RCC_OscInitStruct.OscillatorType = RCC_OSCILLATORTYPE_HSE;
-	RCC_OscInitStruct.HSEState = RCC_HSE_ON;
-	RCC_OscInitStruct.HSIState = RCC_HSI_OFF;
-	RCC_OscInitStruct.PLL.PLLState = RCC_PLL_ON;
-	RCC_OscInitStruct.PLL.PLLSource = RCC_PLLSOURCE_HSE;
-	RCC_OscInitStruct.PLL.PLLM = 25;
-	RCC_OscInitStruct.PLL.PLLN = 432;
-	RCC_OscInitStruct.PLL.PLLP = RCC_PLLP_DIV2;
-	RCC_OscInitStruct.PLL.PLLQ = 9;
-	if (HAL_RCC_OscConfig(&RCC_OscInitStruct) != HAL_OK) {
-		Error_Handler();
-	}
-
-	if (HAL_PWREx_EnableOverDrive() != HAL_OK) {
-		Error_Handler();
-	}
-
-	RCC_ClkInitStruct.ClockType = (RCC_CLOCKTYPE_SYSCLK | RCC_CLOCKTYPE_HCLK
-			| RCC_CLOCKTYPE_PCLK1 | RCC_CLOCKTYPE_PCLK2);
-	RCC_ClkInitStruct.SYSCLKSource = RCC_SYSCLKSOURCE_PLLCLK;
-	RCC_ClkInitStruct.AHBCLKDivider = RCC_SYSCLK_DIV1;
-	RCC_ClkInitStruct.APB1CLKDivider = RCC_HCLK_DIV4;
-	RCC_ClkInitStruct.APB2CLKDivider = RCC_HCLK_DIV2;
-	if (HAL_RCC_ClockConfig(&RCC_ClkInitStruct, FLASH_LATENCY_7) != HAL_OK) {
-		Error_Handler();
-	}
-}
-static void Error_Handler(void) {
-	while (1) {
-	}
-}
-static void MPU_Config(void) {
-	MPU_Region_InitTypeDef MPU_InitStruct;
-
-	HAL_MPU_Disable();
-
-	MPU_InitStruct.Enable = MPU_REGION_ENABLE;
-	MPU_InitStruct.BaseAddress = 0x20010000;
-	MPU_InitStruct.Size = MPU_REGION_SIZE_256KB;
-	MPU_InitStruct.AccessPermission = MPU_REGION_FULL_ACCESS;
-	MPU_InitStruct.IsBufferable = MPU_ACCESS_NOT_BUFFERABLE;
-	MPU_InitStruct.IsCacheable = MPU_ACCESS_CACHEABLE;
-	MPU_InitStruct.IsShareable = MPU_ACCESS_SHAREABLE;
-	MPU_InitStruct.Number = MPU_REGION_NUMBER0;
-	MPU_InitStruct.TypeExtField = MPU_TEX_LEVEL0;
-	MPU_InitStruct.SubRegionDisable = 0x00;
-	MPU_InitStruct.DisableExec = MPU_INSTRUCTION_ACCESS_ENABLE;
-
-	HAL_MPU_ConfigRegion(&MPU_InitStruct);
-
-	HAL_MPU_Enable(MPU_PRIVILEGED_DEFAULT);
-}
-static void CPU_CACHE_Enable(void) {
-	SCB_EnableICache();
-	SCB_EnableDCache();
-}
-
-#ifdef  USE_FULL_ASSERT
-void assert_failed(uint8_t* file, uint32_t line)
-{
-	while (1)
-	{
-	}
-}
-#endif
